# docs/swagger.yaml
definitions:
  models.Post:
    properties:
      content:
        type: string
      id:
        type: integer
      post_images:
        items:
          $ref: '#/definitions/models.PostImage'
        type: array
      published_at:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        type: string
      user_id:
        type: integer
      vote_count:
        type: integer
      votes:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.PostImage:
    properties:
      id:
        type: integer
      post_id:
        type: integer
      url:
        type: string
    type: object
  models.Review:
    properties:
      comment:
        type: string
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Tag:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    type: object
...
  requests.ReviewRequest:
    properties:
      comment:
        type: string
    required:
    - comment
    type: object
...
  results.IDResult:
    properties:
      id:
        type: integer
    type: object
  results.JSONResult:
    properties:
      data: {}
      message:
        type: string
    type: object
  results.ReviewResult:
    properties:
      review:
        $ref: '#/definitions/requests.ReviewRequest'
    type: object
  results.RoleResult:
    properties:
      role:
        $ref: '#/definitions/models.Role'
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /api/v1/posts/{id}/reviews:
    post:
      description: Create a new post review.
      parameters:
      - description: the request body to create a new post review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/requests.ReviewRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: post id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.ReviewResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Create new post review
      tags:
      - Post Reviews
  /api/v1/posts/{id}/reviews/{review_id}:
    delete:
      description: Delete an existing post review by ID.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: review id
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.IDResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Delete post review
      tags:
      - Post Reviews
    put:
      description: Update an existing post review by ID.
      parameters:
      - description: the request body to update a post review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/requests.ReviewRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: review id
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.ReviewResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Update post review
      tags:
      - Post Reviews
...
  /api/v1/roles:
    get:
      description: Get a list of roles, only admin can do this.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/results.RoleResult'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get all roles (admin only)
      tags:
      - Roles
    post:
      description: Create a new role, only admin can do this.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: the request body to create a new role
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.RoleResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Create new role (admin only)
      tags:
      - Roles
  /api/v1/roles/{id}:
    delete:
      description: Delete an existing role by ID, only admin can do this.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: the role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.IDResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete role by ID (admin only)
      tags:
      - Roles
    post:
      description: Create an existing role by ID, only admin can do this.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: the role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.RoleResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get role by ID (admin only)
      tags:
      - Roles
    put:
      description: Update an existing role by ID, only admin can do this.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: the role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.RoleResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Update role by ID (admin only)
      tags:
      - Roles
...
swagger: "2.0"